backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
apscheduler==3.10.4
httpx==0.25.2
pandas==2.1.4
numpy==1.26.2
akshare>=1.17.0
pydantic==2.5.2
python-multipart==0.0.6
greenlet>=3.0.0
python-dotenv>=1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
